add signup functionality and enhance todo management features

+++ server/controller/todo.js
@@ -2,13 +2,14 @@ import { Todo } from "../models/todo.js";
 
 export const createTodo = async (req, res) => {
   try {
+    console.log(req.user);
     const { title, description } = req.body;
     if (!title || !description) {
       return res
         .status(400)
         .json({ success: false, message: "All fields are required" });
     }
-    const todo = await Todo.create({ title, description });
+    const todo = await Todo.create({ title, description, userId: req.user.id });
     return res
       .status(201)
       .json({ success: true, message: "Todo created successfully", todo });
@@ -19,13 +20,15 @@ export const createTodo = async (req, res) => {
 
 export const getTodos = async (req, res) => {
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find({ userId: req.user.id });
     return res.status(200).json({
       success: true,
       todos: todos.length > 0 ? todos : "No todos found",
     });
   } catch (error) {
-    console.log(error);
+    return res
+      .status(500)
+      .json({ success: false, message: "Internal server error" });
   }
 };
 
@@ -33,13 +36,14 @@ export const updateTodo = async (req, res) => {
   try {
     const id = req.params.id;
     const updates = req.body;
+
     if (!updates || Object.keys(updates).length === 0) {
       return res.status(400).json({
         success: false,
         message: "At least one field is required to update",
       });
     }
-    const todo = await Todo.findById(id);
+    const todo = await Todo.findById({ _id: id, userId: req.user.id });
     if (!todo) {
       return res
         .status(404)
@@ -53,19 +57,51 @@ export const updateTodo = async (req, res) => {
       .status(200)
       .json({ success: true, message: "Todo updated successfully", todo });
   } catch (error) {
-    console.log(error);
+    return res
+      .status(500)
+      .json({ success: false, message: "Internal server error" });
   }
 };
 
 export const deleteTodo = async (req, res) => {
   try {
     const id = req.params.id;
-    await Todo.findByIdAndDelete(id);
+    await Todo.findByIdAndDelete({ _id: id, userId: req.user.id });
 
     return res
       .status(200)
       .json({ success: true, message: "Todo deleted successfully" });
   } catch (error) {
-    console.log(error);
+    return res
+      .status(500)
+      .json({ success: false, message: "Internal server error" });
+  }
+};
+
+export const toggleTodoCompletion = async (req, res) => {
+  try {
+    const todo = await Todo.findOne({
+      _id: req.params.id,
+      userId: req.user.id,
+    });
+    if (!todo) {
+      return res
+        .status(404)
+        .json({ success: false, message: "Todo not found" });
+    }
+
+    // Toggle the completion status
+    todo.completed = !todo.completed;
+    await todo.save();
+
+    return res.status(200).json({
+      success: true,
+      message: "Todo completion status updated",
+      todo,
+    });
+  } catch (error) {
+    return res
+      .status(500)
+      .json({ success: false, message: "Internal server error" });
   }
 };
